Move the tinycc code as a plugin ##types
